Commented mouse click movement in Player.
Commented mouse click movement in Player.

+++ OOP Teamwork v1.0/The Powerful Game/Entities/Enemy.cs
@@ -6,7 +6,7 @@ namespace The_Powerful_Game.Entities
 
     public class Enemy : Entity
     {
-        public Enemy(string name, int x, int y, int healthPoints, int armorPoints, int damage, double attackSpeed, Image img)
+        public Enemy(string name, double x, double y, int healthPoints, int armorPoints, int damage, double attackSpeed, Image img)
             : base(name, x, y, healthPoints, armorPoints, damage, attackSpeed, img)
         {
         }
